pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "firebase-admin>=7.1.0",
    "google-cloud>=0.34.0",
    "nltk>=3.9.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentencepiece>=0.2.1",
    "torch>=2.8.0",
    "transformers>=4.55.2",
    "uvicorn>=0.35.0",
]
